+++ addon/components/modal-dialog.js
@@ -1,6 +1,6 @@
 /**
   @module ember-flexberry
- */
+*/
 
 import Ember from 'ember';
 
@@ -23,7 +23,7 @@ import Ember from 'ember';
 
   @class ModalDialog
   @extends <a href="http://emberjs.com/api/classes/Ember.Component.html">Ember.Component</a>
- */
+*/
 export default Ember.Component.extend({
   /**
     Size of Semantic UI modal.
@@ -35,43 +35,43 @@ export default Ember.Component.extend({
     @property sizeClass
     @type String
     @default 'small'
-   */
+  */
   sizeClass: 'small',
 
   /**
-    Flag: indicates whether an image content is viewed at modal dialog or not.
+    Flag indicates whether an image content is viewed at modal dialog or not.
 
     @property viewImageContent
     @type Boolean
     @default false
-   */
+  */
   viewImageContent: false,
 
   /**
-    Flag: indicates whether to show apply button or not.
+    Flag indicates whether to show apply button or not.
 
     @property useOkButton
     @type Boolean
     @default true
-   */
+  */
   useOkButton: true,
 
   /**
-    Flag: indicates whether to show close button or not.
+    Flag indicates whether to show close button or not.
 
     @property useCloseButton
     @type Boolean
     @default true
-   */
+  */
   useCloseButton: true,
 
   /**
-    Flag: indicates toolbar visibility, `true` if at least one of buttons is visible.
+    Flag indicates toolbar visibility, `true` if at least one of buttons is visible.
 
     @property toolbarVisible
     @type Boolean
     @readOnly
-   */
+  */
   toolbarVisible: Ember.computed('useOkButton', 'useCloseButton', function () {
     return this.get('useOkButton') || this.get('useCloseButton');
   }),
@@ -82,12 +82,20 @@ export default Ember.Component.extend({
     @property settings
     @type Object
     @default {}
-   */
+  */
   settings: {},
 
   /**
+    Service that triggers lookup events.
+
+    @property lookupEventsService
+    @type Service
+  */
+  lookupEventsService: Ember.inject.service('lookup-events'),
+
+  /**
     Initializes DOM-related component's logic.
-   */
+  */
   didInsertElement() {
     let _this = this;
     let modalSettings = Ember.$.extend({
@@ -97,17 +105,21 @@ export default Ember.Component.extend({
 
         onApprove: function () {
           _this.sendAction('ok');
+          _this.get('lookupEventsService').showDialogTrigger(false);
         },
         onDeny: function () {
           _this.sendAction('close');
+          _this.get('lookupEventsService').showDialogTrigger(false);
         },
         onHidden: function () {
           _this.sendAction('close');
 
           // IE doesn't support "this.remove()", that's why "Ember.$(this).remove()" is used.
           Ember.$(this).remove();
+          _this.get('lookupEventsService').showDialogTrigger(false);
         },
         onVisible: function () {
+          _this.get('lookupEventsService').showDialogTrigger(true);
           Ember.run(() => {
             _this.sendAction('created', Ember.$(this));
           });
@@ -115,6 +127,6 @@ export default Ember.Component.extend({
       },
       _this.get('settings'));
 
-    this.$('.ui.modal').modal(modalSettings).modal('show');
+    this.$('.ui.modal').modal(modalSettings).modal('show');    
   },
 });