+++ addon/services/lookup-events.js
@@ -1,0 +1,19 @@
+/**
+  @module ember-flexberry
+*/
+
+import Ember from 'ember';
+
+/**
+  Service for triggering lookup events.
+
+  @class LookupEvents
+  @extends Ember.Service
+  @uses Ember.Evented
+  @public
+*/
+export default Ember.Service.extend(Ember.Evented, {
+  showDialogTrigger(modalIsShow) {
+    this.trigger('setModalIsShow', modalIsShow);
+  }
+});