+++ app/services/lookup-events.js
@@ -1,0 +1,2 @@
+import LookupEvents from 'ember-flexberry/services/lookup-events';
+export default LookupEvents;